refactor: app theme applies to the whole app, not just main screen
